kda: export chunked prefill states in fp32 (GLM5 same-class fix as P3-29 GDN)
kda_chunk_delta_h exported per-chunk states h in k.dtype (bf16) while the
recurrence carries fp32 registers and the track snapshots copy h rows into
the fp32 linear-state pool -- every GLM5 KDA snapshot was silently
bf16-rounded, the same root cause as P3-29 on qwen3_5_moe GDN.

- h is now fp32; chunk_gla_fwd_kernel_o casts b_h back to the query dtype
  before tl.dot (the dot already cast at the use site), so prefill outputs
  are bit-identical either way and snapshots gain exact fp32.
- tests/kernels/test_kda.py: chunked-path pins (fp32 h, snapshot-resume
  bit-exact vs cold, rounding sensitivity via fp32 final state). Test
  harness clones inputs per call: chunk_kda_with_fused_gate writes its
  output in-place into v, which otherwise self-contaminates repeat runs
  and looks exactly like a kernel race.

+++ python/freetoken/kernel/fla/kda.py
@@ -780,7 +780,9 @@ def chunk_gla_fwd_kernel_o(
         # [BT, BK]
         b_qg = (b_q * exp2(b_g)).to(b_q.dtype)
         # [BV, BK]
-        b_h = tl.load(p_h, boundary_check=(0, 1))
+        # h is exported in fp32 (see kda_chunk_delta_h); cast back to the
+        # query dtype so tl.dot consumes the same values as a bf16 buffer.
+        b_h = tl.load(p_h, boundary_check=(0, 1)).to(b_q.dtype)
         # [BT, BV]
         if i_k >= 0:
             b_o += tl.dot(b_qg, tl.trans(b_h).to(b_qg.dtype))
